move Packsack.cs to item namespace
